Initial Commit - Currently the bot is very stable
Just finished moving to MySQL rather than a janky json file.
All sout() and some logging are for debugging, and most will be removed when it is complete.
There are a few more additions to be made before this is 100% ready - such as more error handling and better messages.
I need to add checks on NMC to see if the bot returns "success" or "failure" for a few things though. Will do tomorrow